requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
cookies==2.2.1
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.0
Pillow==5.2.0
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
redis==2.10.6
requests==2.19.1
responses==0.9.0
six==1.11.0
tox==3.2.1
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.11